dns-hook-stuff/requirements.txt
cloudflare~=2.3.0
transip~=2.0.0
pyyaml~=5.1.1
requests~=2.22.0
ovh~=0.5.0